more naming changes
its worth talking about this a bit

im somewhat unclear on what the right way to name things is here. in
principle since this package isn't intended to be imported i could just
not export anything, but it's very convenient to name the `Qid` method
as such so i can have a field `qid` that it can just return. If it was
private I'd have to do something goofier.

I'm not sure whether this is correct or reasonable