wip getting truncation to work
need to implement cacheing at every level so that duplicate walks return
the *same object* instead of new in memory copies of it. here's the
story of how we got here:

0. found encoding bug in go-9p and hacked in a fix. need to make actual
fix and submit upstream
1. tried to truncate and sync this to the server. doesn't work becuase
if the content is empty or just whitespace, github deletes the file.
2. tried to use placeholder text to keep the file around. this "works",
but multiple writes in rapid succession trigger a bug in the github API
where what's displayed in the web interface doesn't match what's
displayed in the API
3. tried to just do the truncation locally, on the assumption that
syncing it to the server doesn't matter since a write is probably going
to follow in quick succession.
4. doesn't work because writing the file walks to it again, which
creates a new object in memory so the fact that it was truncated
previously is lost.

so we need to make sure walks are "stable" in terms of the objects in
memory they go to

another creative solution to this problem would be to be willing to sync
truncations to the server, causing the file in question to be deleted,
but be willing to implicitly create it again on writes if we can find
the file in the gist's history. this also is incorrect however, since
there's no way to distinguish truncation from file actually being
deleted